tests: vcs: split off repo creation helpers
test_changesets.py needs to create a repository, and is currently
duplicating some of the base setup code.

Instead, split out the setup_class method to allow reuse.

Note that the repo_path variable previously registered to 'self' is not
actually used by anyone, and moreover is not actually necessary as the path
can be obtained via the repo object too. Therefore, just make it a local
variable.