ini: drop setting initial_repo_scan
The ini setting 'initial_repo_scan' caused a repository scan on each startup
of Kallithea. The accompanying comment already warns that the feature should
be disabled after the first run to improve startup time.

Now that setup-db is performing the initial repository scan, and considering
that administrators can always request a new scan using 'gearbox repo-scan'
or via the web interface, the 'initial_repo_scan' feature is no longer
considered useful and is removed with this commit.