docs: reorganize index page
Reorganize the index page of the manual into three main sections:
administrator guide, user guide, developer guide. In principle,
administrators, users and developers are independent roles that can be
fulfilled by different people.

'Administrators' install, configure and maintain Kallithea. They have
'admin' privilege in the web interface.

'Users' do not have access to the Kallithea installation files. They can
only access the web interface, and generally do not have 'admin' privilege.
They may be the owner of repositories, repository groups, or user groups,
and as such modify the settings of these objects, even though they cannot
modify the settings of Kallithea itself.

'Developers' modify the Kallithea source code. They may or may not
contribute these changes back to the Kallithea community.



The 'Readme' section is a bit outside of this classification. As it serves
as a basic introduction to Kallithea, it is kept first.

The 'Index' and 'Search' functionality are not really 'part of' the manual,
so should not be under 'Other topics'. Instead they are put at the very top
so they are readily visible.


This commit is only changing the index page. Subsequent commits can move and
split some documentation files in the repository to mirror this structure.
For example, the 'Version control systems support' page has information for
administrators (ini settings and importing existing repositories) as well as
for users (handling Mercurial subrepositories).