admin: hooks: restore delete functionality as intended
Commit 9d34bea3059d9abd0d912f37a2475ee67c8e2918 ("style: various minor-ish
markup changes, preparing for Bootstrap") partially broke the delete
functionality of hooks. When clicking the delete button, the hook is deleted
via AJAX and then the corresponding form group is intended to be removed
visually.  This relies on an 'id' attribute on the form-group div.

The mentioned commit moved the div with the 'id' attribute outside the loop
iterating over the different hooks, so that there no longer is a div with
the id expected by the delete button. The hook would still be deleted, but
the page visually still looks the same until refresh/Save.

Move the diff back into the loop. This causes a little more visual
separation between the different hooks, but still looks OK.
The layout of the built-in hooks (that can't be removed) is left untouched.
Making the same change there would be possible but there the extra vertical
whitespace seems somewhat unnecessary, although this is personal.