style: fix comment bubble appearance upon zoom
When browser zoom is used, the font size increases. As a result, the text
balloon icon used in the comment bubble increases as well. However, the
rectangle in which the icon resides is kept the same size, causing the
result to look bad.

Use 'em' units, which are relative to the font size, rather than pixels, for
all relevant style properties of the bubble, including size, padding and
border-radius. With these settings, the bubble always looks well, regardless
of zoom value.