controllers: changeset: always allow status changes
Don't disallow status changes on changesets that are part of a pull request
and on which the last status change happened via the pull request.

This odd restriction was already previously highlighted by Mads Kiilerich as
'RLY?' in commit 7834f845505aec3086f525600c81209a40b495ef, so it seems fair
to remove it.

When dont_allow_on_closed_pull_request no longer is set,
StatusChangeOnClosedPullRequestError will no longer be raised.