front-end: Introduce 'front-end' directory with source files for building the front-end
The top level with package.json is not included when installing with pip, and
'kallithea-cli front-end-build' and npm would thus fail.

Instead, introduce 'kallithea/front-end/'. It is under 'kallithea/' and is thus
included when doing pip install, but it is just a source directory and not
under "public". Most of the "less" stuff should probably move there. And
probably also most things that now are checked in under "public", so a fully
populated public front-end directory can be built anywhere, without write
access to the Python installation directory.