auth: drop UserCreationError handling for AuthUser.from_cookie
Simplify the code a bit, and take an unnecessary concern out of sight.

Users with cookie auth must already exist in the system, and they will thus
never be created on the fly when using a cookie for authentication.

If completely weird things should happen anyway, it would probably be better to
get a clear stack trace.