middleware: drop fix_PATH
Back out ccbdff90e5a0. That seemed like an odd hack. In order to work properly,
it should not only be applied for protocol access middleware, but also for web
UI and for commands. So evidently, it is not really necessary.

The problem it describes is fixed much better in 5e501b6ee639 by setting the
right python executable in the hook scripts, further improved in 1bafb2d07709
and 6df08d78f8e7 to *actually* use the right python executable.