dependencies: remove duplicate dependencies in dev_requirements.txt
Babel and WebOb were present both in setup.py and dev_requirements.txt.
When explicitly setting minimum dependencies in setup.py and running:

     pip install --upgrade -e . -r dev_requirements.txt python-ldap python-pam

this duplication replaced useful diagnostic message from pip with less
useful ones.
For example, the following message was displayed when the Babel dependency
duplication is present:

    kallithea 0.4.0rc1 has requirement Babel<2.7,==0.9.6, but you'll have babel 2.6.0 which is incompatible.

When removing the duplication in dev_requirements.txt, this becomes:

    sphinx 1.7.9 has requirement babel!=2.0,>=1.3, but you'll have babel 0.9.6 which is incompatible.

which makes it clear that to solve this problem, we need to bump the minimum
dependency for Babel in setup.py from 0.9.6 to 1.3.