git: fix handling of submodules that are not in the repo root (Issue #337)
GitChangeset.get_nodes() did not handle submodules correctly if they were
not located in the repository root. The file .gitmodules was searched in the
'tree' object being handled, which would be a subdirectory, while the real
.gitmodules file is in the root tree of the repository.

Instead of using 'name', the 'path' should be used.

This problem was noticed during indexing of such repositories.