.hgtags: remove accidental unused double tag of 0.4.0rc1
Initially, when tagging 0.4.0rc1 I made a mistake locally, fixed it, then
stripped the tagging commit, and retagged.
However, it seems something went wrong in this procedure and the original
commit also is shown in .hgtags.

This commit never got pushed, is hidden (obsolete) in my local repository,
so remove the corresponding line in .hgtags to avoid any confusion.